HSD8-129 Call video plugins to get video title for entity name
