+++ src/Plugin/BundleSuggestion/BundleSuggestionBase.php
@@ -65,4 +65,8 @@ abstract class BundleSuggestionBase extends PluginBase implements BundleSuggesti
       ->loadMultiple($bundles ?: NULL);
   }
 
+  public function getName($input) {
+    return NULL;
+  }
+
 }
