+++ stanford_media.services.yml
@@ -3,13 +3,10 @@ services:
     class: Drupal\stanford_media\Routing\RouteSubscriber
     tags:
       - { name: event_subscriber }
-  stanford_media.bundle_suggestion:
-    class: '\Drupal\stanford_media\Service\BundleSuggestion'
-    arguments: ['@entity_type.manager', '@plugin.manager.bundle_suggestion_manager', '@module_handler']
   plugin.manager.media_embed_dialog_manager:
     class: Drupal\stanford_media\Plugin\MediaEmbedDialogManager
     parent: default_plugin_manager
   plugin.manager.bundle_suggestion_manager:
     class: Drupal\stanford_media\Plugin\BundleSuggestionManager
     parent: default_plugin_manager
-    arguments: ['@entity_field.manager']
+    arguments: ['@entity_field.manager', '@entity_type.manager']
